requirements.txt
streamlit
snowflake-connector-python
pandas
altair
snowflake-snowpark-python